refactor: 인증 예외 정리 및 데이터 접근 예외 처리 위임, Optional 반환
- Optional로 값이 없을 수 있음을 나타낸다
