--- src/main/java/cart/controller/auth/exception/AuthenticateFailException.java
@@ -1,8 +1,0 @@
-package cart.controller.auth.exception;
-
-public class AuthenticateFailException extends AuthenticationException {
-
-    public AuthenticateFailException() {
-        super("잘못된 인증 정보입니다");
-    }
-}
