--- src/main/java/cart/controller/auth/exception/EmptyAuthenticationException.java
@@ -1,8 +1,0 @@
-package cart.controller.auth.exception;
-
-public class EmptyAuthenticationException extends AuthenticationException {
-
-    public EmptyAuthenticationException() {
-        super("인증 정보가 없습니다");
-    }
-}
