+++ src/main/java/cart/controller/auth/exception/InvalidAuthenticationException.java
@@ -3,6 +3,6 @@ package cart.controller.auth.exception;
 public class InvalidAuthenticationException extends AuthenticationException {
 
     public InvalidAuthenticationException() {
-        super("잘못된 인증 정보입니다");
+        super("유효하지 않은 인증 정보입니다");
     }
 }
